tests: refactor (#1587)
