+++ __tests__/e2e/frontmatter/multiple-levels-outline.md
@@ -1,0 +1,41 @@
+---
+title: Multiple Levels Outline
+editLink: true
+outline: 'deep'
+---
+
+# h1 - 1
+
+Lorem ipsum
+
+## h2 - 1
+
+Lorem ipsum
+
+### h3 - 1
+
+Lorem ipsum
+
+#### h4 - 1
+
+Lorem ipsum
+
+### h3 - 2
+
+Lorem ipsum
+
+#### h4 - 2
+
+Lorem ipsum
+
+## h2 - 2
+
+Lorem ipsum
+
+### h3 - 3
+
+Lorem ipsum
+
+#### h4 - 3
+
+Lorem ipsum
