+++ __tests__/e2e/multi-sidebar/index.md
@@ -1,0 +1,1 @@
+# Multi Sidebar
