+++ __tests__/e2e/package.json
@@ -1,0 +1,6 @@
+{
+  "private": true,
+  "devDependencies": {
+    "vitepress": "workspace:*"
+  }
+}
