+++ __tests__/e2e/shims.ts
@@ -1,0 +1,6 @@
+import { type Page } from 'playwright-chromium'
+
+declare global {
+  var page: Page
+  var goto: (path: string) => Promise<void>
+}
