+++ __tests__/e2e/static-data/data.md
@@ -1,0 +1,7 @@
+# Static Data
+
+<script setup lang="ts">
+import { data } from './static.data.js'
+</script>
+
+{{ data }}
