+++ __tests__/e2e/static-data/data/foo.json
