--- __tests__/shims.ts
@@ -1,1 +1,0 @@
-export default '{}'
