+++ __tests__/unit/shims.ts
@@ -1,0 +1,1 @@
+export default {}
