--- __tests__/vitest.config.ts
@@ -1,19 +1,0 @@
-import { dirname, resolve } from 'path'
-import { fileURLToPath } from 'url'
-import { defineConfig } from 'vite'
-
-const dir = dirname(fileURLToPath(import.meta.url))
-
-export default defineConfig({
-  resolve: {
-    alias: {
-      '@siteData': resolve(dir, './shims.ts'),
-      client: resolve(dir, '../src/client'),
-      node: resolve(dir, '../src/node'),
-      vitepress: resolve(dir, '../src/client')
-    }
-  },
-  test: {
-    globals: true
-  }
-})
