--- examples/configured/index.md
@@ -1,3 +1,0 @@
-# Full Configured VitePress Example
-
-WIP
