--- examples/configured/multi-sidebar/index.md
@@ -1,5 +1,0 @@
----
-title: Multi Sidebar Test
----
-
-# Multi Sidebar Test
