--- examples/configured/package.json
@@ -1,13 +1,0 @@
-{
-  "private": true,
-  "type": "module",
-  "name": "vitepress-example-configured",
-  "scripts": {
-    "dev": "vitepress dev",
-    "build": "vitepress build",
-    "serve": "vitepress serve"
-  },
-  "devDependencies": {
-    "vitepress": "workspace:*"
-  }
-}
