--- examples/configured/static-data/data.md
@@ -1,5 +1,0 @@
-<script setup>
-import { data } from './static.data.js'
-</script>
-
-{{ data }}
