--- examples/configured/text-literals/index.md
@@ -1,5 +1,0 @@
----
-title: Text Literals
----
-
-# Text Literals Test
