--- examples/minimal/package.json
@@ -1,12 +1,0 @@
-{
-  "private": true,
-  "name": "vitepress-example-minimal",
-  "scripts": {
-    "dev": "vitepress dev",
-    "build": "vitepress build",
-    "serve": "vitepress serve"
-  },
-  "devDependencies": {
-    "vitepress": "workspace:*"
-  }
-}
