--- examples/test-utils.ts
@@ -1,7 +1,0 @@
-import { page } from './vitestSetup'
-
-export * from './vitestSetup'
-
-export async function waitForLayout() {
-  await page.waitForSelector('#app .Layout')
-}
