--- examples/tsconfig.json
@@ -1,21 +1,0 @@
-{
-  "include": ["."],
-  "exclude": ["**/dist/**"],
-  "compilerOptions": {
-    "target": "ES2020",
-    "module": "ESNext",
-    "outDir": "dist",
-    "baseUrl": ".",
-    "allowJs": true,
-    "esModuleInterop": true,
-    "resolveJsonModule": true,
-    "moduleResolution": "Node",
-    "skipLibCheck": true,
-    "noUnusedLocals": true,
-    "jsx": "preserve",
-    "types": ["vitepress/client", "vitest/globals", "node"],
-    "paths": {
-      "~utils": ["./test-utils.ts"]
-    }
-  }
-}
