+++ package.json
@@ -52,7 +52,7 @@
     "dev": "rimraf dist && run-s dev-shared dev-start",
     "dev-start": "run-p dev-client dev-node dev-watch",
     "dev-client": "tsc -w -p src/client",
-    "dev-node": "DEV=true pnpm run build-node -w",
+    "dev-node": "DEV=true pnpm build-node -w",
     "dev-shared": "node scripts/copyShared",
     "dev-watch": "node scripts/watchAndCopy",
     "build": "run-s build-prepare build-client build-node",
@@ -61,34 +61,31 @@
     "build-node": "tsc -p src/node --noEmit && rollup --config rollup.config.ts --configPlugin esbuild",
     "format": "prettier --check --write .",
     "format-fail": "prettier --check .",
-    "lint": "pnpm run format",
-    "lint-fail": "pnpm run format-fail",
-    "test": "vitest -r __tests__",
-    "test-unit": "vitest run -r __tests__",
-    "test-serve": "vitest run -c vitest.config.e2e.ts",
-    "test-build": "VITE_TEST_BUILD=1 vitest run -c vitest.config.e2e.ts",
-    "debug-serve": "VITE_DEBUG_SERVE=1 vitest run -c vitest.config.e2e.ts",
-    "debug-build": "VITE_TEST_BUILD=1 VITE_PRESERVE_BUILD_ARTIFACTS=1 vitest run -c vitest.config.e2e.ts",
-    "check": "run-s lint test-unit docs-build",
+    "check": "run-s format-fail build test",
+    "test": "run-p --aggregate-output test-unit test-serve test-build",
+    "test-unit": "vitest run -r __tests__/unit",
+    "test-serve": "vitest run -r __tests__/e2e",
+    "test-build": "VITE_TEST_BUILD=1 pnpm test-serve",
+    "debug-serve": "DEBUG=1 vitest -r __tests__/e2e",
+    "debug-build": "VITE_TEST_BUILD=1 pnpm debug-serve",
     "changelog": "conventional-changelog -p angular -i CHANGELOG.md -s",
     "release": "node scripts/release.js",
     "docs": "run-p dev docs-dev",
-    "docs-dev": "wait-on -d 1000 dist/node/cli.js && node ./bin/vitepress dev docs",
+    "docs-dev": "wait-on -d 100 dist/node/cli.js && node ./bin/vitepress dev docs",
     "docs-debug": "node --inspect-brk ./bin/vitepress dev docs",
     "docs-build": "run-s build docs-build-only",
     "docs-build-only": "node ./bin/vitepress build docs",
-    "docs-serve": "node ./bin/vitepress serve docs",
-    "ci-docs": "run-s docs-build"
+    "docs-serve": "node ./bin/vitepress serve docs"
   },
   "dependencies": {
     "@docsearch/css": "^3.3.0",
     "@docsearch/js": "^3.3.0",
-    "@vitejs/plugin-vue": "^3.1.2",
+    "@vitejs/plugin-vue": "^3.2.0",
     "@vue/devtools-api": "^6.4.5",
     "@vueuse/core": "^9.4.0",
     "body-scroll-lock": "4.0.0-beta.0",
     "shiki": "^0.11.1",
-    "vite": "^3.1.8",
+    "vite": "^3.2.2",
     "vue": "^3.2.41"
   },
   "devDependencies": {
@@ -118,7 +115,7 @@
     "@types/markdown-it-emoji": "^2.0.2",
     "@types/micromatch": "^4.0.2",
     "@types/minimist": "^1.2.2",
-    "@types/node": "^18.11.5",
+    "@types/node": "^18.11.9",
     "@types/prompts": "^2.4.1",
     "chokidar": "^3.5.3",
     "compression": "^1.7.4",
@@ -126,13 +123,14 @@
     "cross-spawn": "^7.0.3",
     "debug": "^4.3.4",
     "enquirer": "^2.3.6",
-    "esbuild": "^0.15.12",
+    "esbuild": "^0.15.13",
     "escape-html": "^1.0.3",
     "execa": "^6.1.0",
     "fast-glob": "^3.2.12",
     "fs-extra": "^10.1.0",
+    "get-port": "^5.1.1",
     "lint-staged": "^13.0.3",
-    "lru-cache": "^7.14.0",
+    "lru-cache": "^7.14.1",
     "markdown-it": "^13.0.1",
     "markdown-it-anchor": "^8.6.5",
     "markdown-it-attrs": "^4.1.4",
@@ -154,12 +152,12 @@
     "rollup-plugin-dts": "^4.2.3",
     "rollup-plugin-esbuild": "^4.10.1",
     "semver": "^7.3.8",
+    "shiki-processor": "^0.1.1",
     "simple-git-hooks": "^2.8.1",
     "sirv": "^2.0.2",
-    "shiki-processor": "^0.1.1",
     "supports-color": "^9.2.3",
     "typescript": "~4.8.4",
-    "vitest": "^0.24.3",
+    "vitest": "^0.24.5",
     "vue-tsc": "^1.0.9",
     "wait-on": "^6.0.1"
   },
